requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alabaster==0.7.10
argh==0.26.2              # via watchdog
arrow==0.13.2             # via jinja2-time, pytest-cookies
aspy.yaml==1.3.0          # via pre-commit
astor==0.8.0              # via wemake-python-styleguide
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via flake8-bugbear, flake8-eradicate, packaging, pytest, wemake-python-styleguide
bandit==1.6.2             # via flake8-bandit
binaryornot==0.4.4        # via cookiecutter
certifi==2019.6.16        # via requests
cfgv==2.0.1               # via pre-commit
chardet==3.0.4            # via binaryornot, requests
click==7.0                # via cookiecutter, pip-tools
cookiecutter==1.6.0
decopatch==1.4.5          # via pytest-cases
docutils==0.15.2          # via restructuredtext-lint
entrypoints==0.3          # via flake8
eradicate==1.0            # via flake8-eradicate
flake8-annotations-complexity==0.0.2  # via wemake-python-styleguide
flake8-bandit==2.1.1      # via wemake-python-styleguide
flake8-broken-line==0.1.0  # via wemake-python-styleguide
flake8-bugbear==19.3.0    # via wemake-python-styleguide
flake8-builtins==1.4.1    # via wemake-python-styleguide
flake8-coding==1.3.2      # via wemake-python-styleguide
flake8-commas==2.0.0      # via wemake-python-styleguide
flake8-comprehensions==2.1.0  # via wemake-python-styleguide
flake8-debugger==3.1.0    # via wemake-python-styleguide
flake8-docstrings==1.3.1  # via wemake-python-styleguide
flake8-eradicate==0.2.1   # via wemake-python-styleguide
flake8-executable==2.0.3  # via wemake-python-styleguide
flake8-isort==2.7.0       # via wemake-python-styleguide
flake8-logging-format==0.6.0  # via wemake-python-styleguide
flake8-pep3101==1.2.1     # via wemake-python-styleguide
flake8-polyfill==1.0.2    # via flake8-bandit, flake8-docstrings, pep8-naming
flake8-print==3.1.0       # via wemake-python-styleguide
flake8-quotes==2.1.0      # via wemake-python-styleguide
flake8-rst-docstrings==0.0.10  # via wemake-python-styleguide
flake8-string-format==0.2.3  # via wemake-python-styleguide
flake8==3.7.8             # via flake8-bandit, flake8-broken-line, flake8-bugbear, flake8-builtins, flake8-coding, flake8-commas, flake8-comprehensions, flake8-debugger, flake8-docstrings, flake8-eradicate, flake8-executable, flake8-isort, flake8-pep3101, flake8-polyfill, flake8-print, flake8-quotes, flake8-rst-docstrings, flake8-string-format, wemake-python-styleguide, yesqa
future==0.17.1            # via cookiecutter
gitdb2==2.0.5             # via gitpython
gitpython==2.1.13         # via bandit
identify==1.4.5           # via pre-commit
idna==2.8                 # via requests
importlib-metadata==0.19  # via pluggy, pre-commit, pytest
importlib-resources==1.0.2  # via pre-commit
isort==4.3.21             # via flake8-isort
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.10.1            # via cookiecutter, jinja2-time
makefun==1.6.10           # via decopatch, pytest-cases
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==7.2.0     # via pytest
nodeenv==1.3.3            # via pre-commit
packaging==19.1           # via pytest
pathspec==0.5.9           # via yamllint
pathtools==0.1.2          # via watchdog
pbr==5.4.2                # via stevedore
pep8-naming==0.8.2        # via wemake-python-styleguide
pip-tools==4.0.0
pluggy==0.12.0            # via pytest, tox
poyo==0.5.0               # via cookiecutter
pre-commit==1.18.0
py==1.8.0                 # via pytest, tox
pycodestyle==2.5.0        # via flake8, flake8-bandit, flake8-debugger, flake8-print
pydocstyle==3.0.0         # via flake8-docstrings, wemake-python-styleguide
pyflakes==2.1.1           # via flake8
pygments==2.4.2
pyhamcrest==1.9.0
pyparsing==2.4.2          # via packaging
pytest-cases==1.11.1
pytest-cookies==0.4.0
pytest==4.6.5
python-dateutil==2.8.0    # via arrow
pyyaml==5.1.2             # via aspy.yaml, bandit, pre-commit, watchdog, yamllint
requests==2.22.0          # via cookiecutter
restructuredtext-lint==1.3.0
sh==1.12.14
six==1.12.0               # via bandit, cfgv, flake8-print, makefun, packaging, pip-tools, pre-commit, pydocstyle, pyhamcrest, pytest, pytest-cases, python-dateutil, stevedore, tox
smmap2==2.0.5             # via gitdb2
snowballstemmer==1.9.0    # via pydocstyle
stevedore==1.30.1         # via bandit
testfixtures==6.10.0      # via flake8-isort
tokenize-rt==3.2.0        # via yesqa
toml==0.10.0              # via pre-commit
tox==2.9.1
typing-extensions==3.7.4  # via wemake-python-styleguide
urllib3==1.25.3           # via requests
virtualenv==16.7.2        # via pre-commit, tox
watchdog==0.8.3
wcwidth==0.1.7            # via pytest
wemake-python-styleguide==0.11.1
whichcraft==0.6.0         # via cookiecutter
wrapt==1.11.2             # via pytest-cases
yamllint==1.16.0
yesqa==0.0.11
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via flake8-annotations-complexity, pyhamcrest
